+++ R/save_rds_archive.R
@@ -8,10 +8,10 @@
 #'   scripts) will continue to work while an archived copy of the - otherwise
 #'   overwritten - file will be kept.
 #'
-#'   Please note: If the file does \emph{not} already exist (i.e., if there is nothing to
-#'   overwrite or archive), regular \code{\link[base]{saveRDS}} behavior will be invoked.
-#'   In such a case, all arguments except \code{object} and \code{file} will be
-#'   ignored!
+#'   Please note: If the file does \emph{not} already exist (i.e., if there is
+#'   nothing to overwrite or archive), regular \code{\link[base]{saveRDS}}
+#'   behavior will be invoked. In such a case, all arguments except
+#'   \code{object} and \code{file} will be ignored!
 #'
 #' @param object Object to be saved
 #' @param file Name of the file (path) where the R object is saved to. Note that
@@ -19,12 +19,12 @@
 #' @param archive Logical - should the file be archived if it already exists
 #'   (default), or should it be overwritten (regular saveRDS behavior)?
 #' @param last_modified Logical - should the file name of the archived file be
-#'   appended with the \emph{"last modified"} date/time of the original RDS instead of
-#'   the \emph{current} date/time? Defaults to \code{FALSE}.
+#'   appended with the \emph{"last modified"} date/time of the original RDS
+#'   instead of the \emph{current} date/time? Defaults to \code{FALSE}.
 #' @param with_time Logical - should the file be archived with just a date
 #'   suffix (default) or with a date \bold{and} time suffix? Applies to both
-#'   archiving and modification date. Set to \code{TRUE} if you want to keep several
-#'   versions of files archived on a single day. See details.
+#'   archiving and modification date. Set to \code{TRUE} if you want to keep
+#'   several versions of files archived on a single day. See details.
 #' @param archive_dir_path Character - if desired, path to a dedicated archive
 #'   (sub-)directory (relative to the directory of \code{file}) where the
 #'   archived file will be saved. Will be created if it does not yet exist.
@@ -71,51 +71,54 @@
 #'
 #' @export
 #'
-
 save_rds_archive <- function(object,
                              file = "",
                              archive = TRUE,
                              last_modified = FALSE,
                              with_time = FALSE,
                              archive_dir_path = NULL,
-                             ...){
+                             ...) {
 
-  if (file == "" | !"character" %in% class(file)){
-    stop("'file' must be a non-empty character string")}
+  if (file == "" | !"character" %in% class(file)) {
+    stop("'file' must be a non-empty character string")
+  }
 
-  if (!is.null(archive_dir_path) && archive_dir_path == "")
+  if (!is.null(archive_dir_path) && archive_dir_path == "") {
     stop("must supply a directory name to 'archive_dir_path' if not NULL")
+  }
 
-  if (!is.logical(archive)){
+  if (!is.logical(archive)) {
     archive <- TRUE
     warning("'archive' is not set to a boolean - will use default: ", archive)
   }
 
-  if (!is.logical(with_time)){
+  if (!is.logical(with_time)) {
     with_time <- FALSE
-    warning("'with_time' is not set to a boolean - will use default: ", with_time)
+    warning("'with_time' is not set to a boolean - will use default: ",
+            with_time)
   }
 
   # IF ARCHIVE == TRUE --------------------------------------------------------
 
-  if(archive){
+  if (archive) {
 
     # check if file exists
-    if (file.exists(file)){
+    if (file.exists(file)) {
 
       archived_file <- create_archived_file(file = file,
                                             last_modified = last_modified,
                                             with_time = with_time)
 
-      if (!is.null(archive_dir_path)){
+      if (!is.null(archive_dir_path)) {
 
         # get parent directory
         dname <- dirname(file)
 
         # create archive dir if it does not already exist
-        if(!dir.exists(file.path(dname, archive_dir_path))){
+        if (!dir.exists(file.path(dname, archive_dir_path))) {
           dir.create(file.path(dname, archive_dir_path), recursive = TRUE)
-          message("Created missing archive directory ", sQuote(archive_dir_path))
+          message("Created missing archive directory ",
+                  sQuote(archive_dir_path))
         }
 
         # change path of archived file into 'archive' folder
@@ -126,9 +129,10 @@ save_rds_archive <- function(object,
         # copy (rather than rename) file
         # rename sometimes does not work if the directory itself is changed
         # save return value of the file.copy function
-        # set "overwrite" to TRUE so an existing copy is overwritten (see details)
+        # set "overwrite" to TRUE so an existing copy is overwritten
+        # (see details)
 
-        if (file.exists(archived_file)){
+        if (file.exists(archived_file)) {
           warning("Archived copy already exists - will overwrite!")
         }
 
@@ -138,7 +142,7 @@ save_rds_archive <- function(object,
 
       } else {
 
-        if (file.exists(archived_file)){
+        if (file.exists(archived_file)) {
           warning("Archived copy already exists - will overwrite!")
         }
 
@@ -151,10 +155,10 @@ save_rds_archive <- function(object,
       }
 
       # check return value and if archived file exists
-      if (temp & file.exists(archived_file)){
-
+      if (temp & file.exists(archived_file)) {
         # then save new file under specified name
-        saveRDS(object = object, file = file, ...)}
+        saveRDS(object = object, file = file, ...)
+      }
 
 
     } else {
@@ -170,7 +174,7 @@ save_rds_archive <- function(object,
 
   } else {
 
-    # OTHERWISE USE DEFAULT RDS -------------------------------------------------
+    # OTHERWISE USE DEFAULT RDS -----------------------------------------------
 
     warning("Nothing to overwrite - will use regular saveRDS() behavior. ",
             "Additional arguments will be ignored!")