+++ tests/testthat/test-statusbar.R
@@ -1,10 +1,10 @@
 context("check statusbar")
 
 test_that("error message", {
-  expect_error(statusbar(run = c(1, 2), max.run = 20, percent.max = 10L),
+  expect_error(statusbar(run = c(1, 2), max.run = 20, width = 10L),
                "run needs to be of length one!")
 
-  expect_error(statusbar(run = 1, max.run = c(), percent.max = 10L),
+  expect_error(statusbar(run = 1, max.run = c(), width = 10L),
                "max.run has length 0")
 
 
@@ -12,12 +12,12 @@ test_that("error message", {
 
 
 test_that("example output is right", {
-  expect_output(statusbar(run = 2, max.run = 20, percent.max = 10L),
+  expect_output(statusbar(run = 2, max.run = 20, width = 10L),
                 "\\n \\[=         \\]   10\\.00% - 2")
 
-  expect_output(statusbar(run = "b", max.run = letters[1:3], percent.max = 10L),
+  expect_output(statusbar(run = "b", max.run = letters[1:3], width = 10L),
                 "\\n \\[=======   \\]   66\\.67% - b")
 
-  expect_output(statusbar(run = "b", max.run = letters[2], percent.max = 10L),
+  expect_output(statusbar(run = "b", max.run = letters[2], width = 10L),
                 "\\n \\[==========\\]  100\\.00% - b")
 })